auto extract titles
